requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
sqlalchemy==2.0.36
openai>=1.56.0
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
pydantic==2.9.2
